Split log module
The log module is now split within multiple files for handlers, filters
and formatters. This is supposed to avoid dependency loops while having
code at the most appropriate place.
